Network for each coin and derived keys
network for each coin and derived keys
