+++ DSW.HDWallet/Domain/Coins/CoinType.cs
@@ -1,0 +1,23 @@
+﻿namespace DSW.HDWallet.Domain.Coins
+{
+    public enum CoinType
+    {
+        AZR,
+        BECN,
+        BIR,
+        CFL,
+        SAGA,
+        DASHD,
+        ESK,
+        FLS,
+        _777,
+        KYAN,
+        MOBIC,
+        MONK,
+        OWO,
+        PNY,
+        SAPP,
+        SUV,
+        UCR
+    }
+}
