+++ DSW.HDWallet/Domain/Wallets/IWalletRepository.cs
@@ -1,12 +1,13 @@
-﻿using NBitcoin;
+﻿using DSW.HDWallet.Domain.Coins;
+using NBitcoin;
 
 namespace DSW.HDWallet.Domain.Wallets
 {
     public interface IWalletRepository
     {
-        Wallet Create(Mnemonic mnemonic);
-        Wallet CreateWithPassword(Mnemonic mnemonic, string? password = null);
-        BitcoinAddress Recover(Mnemonic mnemo, string? password = null);
-        BitcoinExtKey CreateDeriveKey(ExtKey masterKey, KeyPath keyPath);
+        Wallet Create(Mnemonic mnemonic, CoinType coinType);
+        Wallet CreateWithPassword(CoinType coinType, Mnemonic mnemonic, string? password = null);
+        BitcoinAddress Recover(CoinType coinType, Mnemonic mnemo, string? password = null);
+        BitcoinExtKey CreateDeriveKey(CoinType coinType, ExtKey masterKey, KeyPath keyPath);
     }
 }
