+++ DSW.HDWallet/Infrastructure/Coins/Azzure.cs
@@ -4,40 +4,40 @@ using NBitcoin.Crypto;
 using NBitcoin.DataEncoders;
 using NBitcoin.Protocol;
 
-namespace DSW.HDWallet.AltCoins
+namespace DSW.HDWallet.Infrastructure.Coins
 {
 
-    public class Kyanite : NetworkSetBase
+    public class Azzure : NetworkSetBase
     {
-        public static Kyanite Instance { get; } = new Kyanite();
+        public static Azzure Instance { get; } = new Azzure();
 
-        public override string CryptoCode => "KYAN";
+        public override string CryptoCode => "AZR";
 
-        private Kyanite()
+        private Azzure()
         {
         }
 
-        public class KyaniteConsensusFactory : ConsensusFactory
+        public class AzzureConsensusFactory : ConsensusFactory
         {
-            private KyaniteConsensusFactory()
+            private AzzureConsensusFactory()
             {
             }
 
-            public static KyaniteConsensusFactory Instance { get; } = new KyaniteConsensusFactory();
+            public static AzzureConsensusFactory Instance { get; } = new AzzureConsensusFactory();
 
             public override BlockHeader CreateBlockHeader()
             {
-                return new KyaniteBlockHeader();
+                return new AzzureBlockHeader();
             }
 
             public override Block CreateBlock()
             {
-                return new KyaniteBlock(new KyaniteBlockHeader());
+                return new AzzureBlock(new AzzureBlockHeader());
             }
         }
 
 #pragma warning disable CS0618 // Type or member is obsolete
-        public class KyaniteBlockHeader : BlockHeader
+        public class AzzureBlockHeader : BlockHeader
         {
             private static byte[] CalculateHash(byte[] data, int offset, int count)
             {
@@ -52,9 +52,9 @@ namespace DSW.HDWallet.AltCoins
             }
         }
 
-        public class KyaniteBlock : Block
+        public class AzzureBlock : Block
         {
-            public KyaniteBlock(KyaniteBlockHeader h) : base(h)
+            public AzzureBlock(AzzureBlockHeader h) : base(h)
             {
             }
 
@@ -67,7 +67,7 @@ namespace DSW.HDWallet.AltCoins
 
         protected override void PostInit()
         {
-            RegisterDefaultCookiePath("KYAN");
+            RegisterDefaultCookiePath("AZR");
         }
 
         protected override NetworkBuilder CreateMainnet()
@@ -79,7 +79,7 @@ namespace DSW.HDWallet.AltCoins
                 MajorityEnforceBlockUpgrade = 8100,
                 MajorityRejectBlockOutdated = 10260,
                 MajorityWindow = 10800,
-                BIP34Hash = new uint256("00000551e93eb0749d40dfafd54b092e78d6612b47bd40de8d099818f65f53c1"),
+                BIP34Hash = new uint256("00000c76e5393b390e03a0c0ecd6e773a99bb906d30aa8bc2a1108146bc78561"),
                 PowLimit = new Target(0 >> 1),
                 MinimumChainWork = new uint256("000000000000000000000000000000000000000000000000c2ba8ca4fb1f06cb"),
                 PowTargetTimespan = TimeSpan.FromSeconds(24 * 60 * 60),
@@ -89,28 +89,28 @@ namespace DSW.HDWallet.AltCoins
                 PowNoRetargeting = false,
                 RuleChangeActivationThreshold = 1916,
                 MinerConfirmationWindow = 2016,
-                ConsensusFactory = KyaniteConsensusFactory.Instance,
+                ConsensusFactory = AzzureConsensusFactory.Instance,
                 SupportSegwit = false,
-                CoinType = 834
+                CoinType = 835
             })
-                .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 46 })
-                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 16 })
-                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 43 })
-                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x88, 0xB2, 0x1E })
-                .SetBase58Bytes(Base58Type.EXT_SECRET_KEY, new byte[] { 0x04, 0x88, 0xAD, 0xE4 })
-                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("Kyanite"))
-                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("Kyanite"))
+                .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 23 })
+                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 24 })
+                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 198 })
+                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x02, 0x2D, 0x25, 0x33 })
+                .SetBase58Bytes(Base58Type.EXT_SECRET_KEY, new byte[] { 0x02, 0x21, 0x31, 0x2B })
+                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("Azzure"))
+                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("Azzure"))
                 .SetMagic(0x191643a0)
-                .SetPort(7757)
-                .SetRPCPort(7758)
+                .SetPort(14725)
+                .SetRPCPort(14724)
                 .SetMaxP2PVersion(70920)
-                .SetName("Kyanite-main")
-                .AddAlias("Kyanite-mainnet")
+                .SetName("Azzure-main")
+                .AddAlias("Azzure-mainnet")
                 .AddDNSSeeds(new[]
                 {
-                    new DNSSeedData("seed1", "seed1.kyancoin.net"),
-                    new DNSSeedData("seed2", "seed2.kyancoin.net"),
-                    new DNSSeedData("seed3", "seed3.kyancoin.net")
+                    new DNSSeedData("seed1", "seed1.azzurecoin.net"),
+                    new DNSSeedData("seed2", "seed2.azzurecoin.net"),
+                    new DNSSeedData("seed3", "seed3.azzurecoin.net")
                 })
                 .AddSeeds(new NetworkAddress[0])
                 .SetGenesis("01000000000000000000000000000000000000000000000000000000000000000000000014e427b75837280517873799a954e87b8b0484f3f1df927888a0ff4fd3a0c9f7bb2eac56f0ff0f1edfa624000101000000010000000000000000000000000000000000000000000000000000000000000000ffffffff8604ffff001d01044c7d323031372d30392d32312032323a30313a3034203a20426974636f696e20426c6f636b204861736820666f722048656967687420343836333832203a2030303030303030303030303030303030303039326431356535623365366538323639333938613834613630616535613264626434653766343331313939643033ffffffff0100ba1dd205000000434104c10e83b2703ccf322f7dbd62dd5855ac7c10bd055814ce121ba32607d573b8810c02c0582aed05b4deb9c4b77b26d92428c61256cd42774babea0a073b2ed0c9ac00000000");
@@ -137,23 +137,23 @@ namespace DSW.HDWallet.AltCoins
                 PowNoRetargeting = false,
                 RuleChangeActivationThreshold = 1512,
                 MinerConfirmationWindow = 2016,
-                ConsensusFactory = KyaniteConsensusFactory.Instance,
+                ConsensusFactory = AzzureConsensusFactory.Instance,
                 SupportSegwit = false,
                 CoinType = 1
             })
-               .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 46 })
-                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 16 })
-                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 43 })
-                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x88, 0xB2, 0x1E })
-                .SetBase58Bytes(Base58Type.EXT_SECRET_KEY, new byte[] { 0x04, 0x88, 0xAD, 0xE4 })
-                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("Kyanite"))
-                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("Kyanite"))
+                .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 23 })
+                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 24 })
+                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 198 })
+                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x02, 0x2D, 0x25, 0x33 })
+                .SetBase58Bytes(Base58Type.EXT_SECRET_KEY, new byte[] { 0x02, 0x21, 0x31, 0x2B })
+                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("Azzure"))
+                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("Azzure"))
                 .SetMagic(0x191643a0)
-                .SetPort(7757)
-                .SetRPCPort(7758)
+                .SetPort(14725)
+                .SetRPCPort(14724)
                 .SetMaxP2PVersion(70920)
-                .SetName("Kyanite-test")
-                .AddAlias("Kyanite-testnet")
+                .SetName("Azzure-test")
+                .AddAlias("Azzure-testnet")
                 .AddSeeds(new NetworkAddress[0])
                 //testnet down for now
                 .SetGenesis("0100000000000000000000000000000000000000000000000000000000000000000000008c5b00d67050180b3a90addb9cd1aabbb3dd79ce20fc071d428ce374581b3f7cde30df5cf0ff0f1e1a1754000101000000010000000000000000000000000000000000000000000000000000000000000000ffffffff3b04ffff001d010433446f676543617368205265706f7765726564204c61756e6368202d20616b736861796e65787573202d204c6971756964333639ffffffff0100000000000000004341047a7df379bd5e6b93b164968c10fcbb141ecb3c6dc1a5e181c2a62328405cf82311dd5b40bf45430320a4f30add05c8e3e16dd56c52d65f7abe475189564bf2b1ac00000000");
@@ -180,22 +180,22 @@ namespace DSW.HDWallet.AltCoins
                 PowNoRetargeting = true,
                 RuleChangeActivationThreshold = 1916,
                 MinerConfirmationWindow = 2016,
-                ConsensusFactory = KyaniteConsensusFactory.Instance,
+                ConsensusFactory = AzzureConsensusFactory.Instance,
                 SupportSegwit = false
             })
-               .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 46 })
-                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 16 })
-                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 43 })
-                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x88, 0xB2, 0x1E })
-                .SetBase58Bytes(Base58Type.EXT_SECRET_KEY, new byte[] { 0x04, 0x88, 0xAD, 0xE4 })
-                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("Kyanite"))
-                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("Kyanite"))
+               .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 23 })
+                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 24 })
+                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 198 })
+                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x02, 0x2D, 0x25, 0x33 })
+                .SetBase58Bytes(Base58Type.EXT_SECRET_KEY, new byte[] { 0x02, 0x21, 0x31, 0x2B })
+                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("Azzure"))
+                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("Azzure"))
                 .SetMagic(0x191643a0)
-                .SetPort(7757)
-                .SetRPCPort(7758)
+                .SetPort(14725)
+                .SetRPCPort(14724)
                 .SetMaxP2PVersion(70920)
-                .SetName("Kyanite-reg")
-                .AddAlias("Kyanite-regtest")
+                .SetName("Azzure-reg")
+                .AddAlias("Azzure-regtest")
                 .AddDNSSeeds(new DNSSeedData[0])
                 .AddSeeds(new NetworkAddress[0])
                 //No regtest at the moment
