+++ DSW.HDWallet/Infrastructure/Coins/Flits.cs
@@ -4,40 +4,40 @@ using NBitcoin.Crypto;
 using NBitcoin.DataEncoders;
 using NBitcoin.Protocol;
 
-namespace DSW.HDWallet.AltCoins
+namespace DSW.HDWallet.Infrastructure.Coins
 {
 
-    public class PEPSCoin : NetworkSetBase
+    public class Flits : NetworkSetBase
     {
-        public static PEPSCoin Instance { get; } = new PEPSCoin();
+        public static Flits Instance { get; } = new Flits();
 
-        public override string CryptoCode => "PEPS";
+        public override string CryptoCode => "FLS";
 
-        private PEPSCoin()
+        private Flits()
         {
         }
 
-        public class PEPSCoinConsensusFactory : ConsensusFactory
+        public class FlitsConsensusFactory : ConsensusFactory
         {
-            private PEPSCoinConsensusFactory()
+            private FlitsConsensusFactory()
             {
             }
 
-            public static PEPSCoinConsensusFactory Instance { get; } = new PEPSCoinConsensusFactory();
+            public static FlitsConsensusFactory Instance { get; } = new FlitsConsensusFactory();
 
             public override BlockHeader CreateBlockHeader()
             {
-                return new PEPSCoinBlockHeader();
+                return new FlitsBlockHeader();
             }
 
             public override Block CreateBlock()
             {
-                return new PEPSCoinBlock(new PEPSCoinBlockHeader());
+                return new FlitsBlock(new FlitsBlockHeader());
             }
         }
 
 #pragma warning disable CS0618 // Type or member is obsolete
-        public class PEPSCoinBlockHeader : BlockHeader
+        public class FlitsBlockHeader : BlockHeader
         {
             private static byte[] CalculateHash(byte[] data, int offset, int count)
             {
@@ -52,9 +52,9 @@ namespace DSW.HDWallet.AltCoins
             }
         }
 
-        public class PEPSCoinBlock : Block
+        public class FlitsBlock : Block
         {
-            public PEPSCoinBlock(PEPSCoinBlockHeader h) : base(h)
+            public FlitsBlock(FlitsBlockHeader h) : base(h)
             {
             }
 
@@ -67,7 +67,7 @@ namespace DSW.HDWallet.AltCoins
 
         protected override void PostInit()
         {
-            RegisterDefaultCookiePath("PEPS");
+            RegisterDefaultCookiePath("FLS");
         }
 
         protected override NetworkBuilder CreateMainnet()
@@ -85,32 +85,32 @@ namespace DSW.HDWallet.AltCoins
                 PowTargetTimespan = TimeSpan.FromSeconds(24 * 60 * 60),
                 PowTargetSpacing = TimeSpan.FromSeconds(1 * 60),
                 PowAllowMinDifficultyBlocks = true,
-                CoinbaseMaturity = 76,
+                CoinbaseMaturity = 100,
                 PowNoRetargeting = false,
                 RuleChangeActivationThreshold = 1916,
                 MinerConfirmationWindow = 2016,
-                ConsensusFactory = PEPSCoinConsensusFactory.Instance,
+                ConsensusFactory = FlitsConsensusFactory.Instance,
                 SupportSegwit = false,
-                CoinType = 847
+                CoinType = 850
             })
-                .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 55 })
-                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 92 })
-                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 117 })
-                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x04, 0x2e, 0x21, 0x71 })
-                .SetBase58Bytes(Base58Type.EXT_SECRET_KEY, new byte[] { 0x08, 0x25, 0x39, 0x2F })
-                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("PEPSCoin"))
-                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("PEPSCoin"))
+                .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 36 })
+                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 48 })
+                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 66 })
+                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x02, 0x2D, 0x25, 0x33 })
+                .SetBase58Bytes(Base58Type.EXT_SECRET_KEY, new byte[] { 0x02, 0x21, 0x31, 0x2B })
+                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("Flits"))
+                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("Flits"))
                 .SetMagic(0x191643a0)
-                .SetPort(15252)
-                .SetRPCPort(16262)
+                .SetPort(32972)
+                .SetRPCPort(32973)
                 .SetMaxP2PVersion(70920)
-                .SetName("PEPSCoin-main")
-                .AddAlias("PEPSCoin-mainnet")
+                .SetName("Flits-main")
+                .AddAlias("Flits-mainnet")
                 .AddDNSSeeds(new[]
                 {
-                    new DNSSeedData("seed1", "seed1.peps.today"),
-                    new DNSSeedData("seed2", "seed2.peps.today"),
-                    new DNSSeedData("seed3", "seed3.peps.today")
+                    new DNSSeedData("seed1", "seed1.flitswallet.app"),
+                    new DNSSeedData("seed2", "seed2.flitswallet.app"),
+                    new DNSSeedData("seed3", "seed3.flitswallet.app")
                 })
                 .AddSeeds(new NetworkAddress[0])
                 .SetGenesis("01000000000000000000000000000000000000000000000000000000000000000000000014e427b75837280517873799a954e87b8b0484f3f1df927888a0ff4fd3a0c9f7bb2eac56f0ff0f1edfa624000101000000010000000000000000000000000000000000000000000000000000000000000000ffffffff8604ffff001d01044c7d323031372d30392d32312032323a30313a3034203a20426974636f696e20426c6f636b204861736820666f722048656967687420343836333832203a2030303030303030303030303030303030303039326431356535623365366538323639333938613834613630616535613264626434653766343331313939643033ffffffff0100ba1dd205000000434104c10e83b2703ccf322f7dbd62dd5855ac7c10bd055814ce121ba32607d573b8810c02c0582aed05b4deb9c4b77b26d92428c61256cd42774babea0a073b2ed0c9ac00000000");
@@ -137,22 +137,23 @@ namespace DSW.HDWallet.AltCoins
                 PowNoRetargeting = false,
                 RuleChangeActivationThreshold = 1512,
                 MinerConfirmationWindow = 2016,
-                ConsensusFactory = PEPSCoinConsensusFactory.Instance,
+                ConsensusFactory = FlitsConsensusFactory.Instance,
                 SupportSegwit = false,
                 CoinType = 1
             })
-               .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 55 })
-                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 92 })
-                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 117 })
-                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x04, 0x2e, 0x21, 0x71 })
-                .SetBase58Bytes(Base58Type.EXT_SECRET_KEY, new byte[] { 0x08, 0x25, 0x39, 0x2F })
-                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("PEPSCoin"))
-                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("PEPSCoin"))
+                .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 36 })
+                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 48 })
+                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 66 })
+                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x02, 0x2D, 0x25, 0x33 })
+                .SetBase58Bytes(Base58Type.EXT_SECRET_KEY, new byte[] { 0x02, 0x21, 0x31, 0x2B })
+                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("Flits"))
+                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("Flits"))
                 .SetMagic(0x191643a0)
-                .SetPort(15252)
+                .SetPort(32972)
+                .SetRPCPort(32973)
                 .SetMaxP2PVersion(70920)
-                .SetName("PEPSCoin-test")
-                .AddAlias("PEPSCoin-testnet")
+                .SetName("Flits-test")
+                .AddAlias("Flits-testnet")
                 .AddSeeds(new NetworkAddress[0])
                 //testnet down for now
                 .SetGenesis("0100000000000000000000000000000000000000000000000000000000000000000000008c5b00d67050180b3a90addb9cd1aabbb3dd79ce20fc071d428ce374581b3f7cde30df5cf0ff0f1e1a1754000101000000010000000000000000000000000000000000000000000000000000000000000000ffffffff3b04ffff001d010433446f676543617368205265706f7765726564204c61756e6368202d20616b736861796e65787573202d204c6971756964333639ffffffff0100000000000000004341047a7df379bd5e6b93b164968c10fcbb141ecb3c6dc1a5e181c2a62328405cf82311dd5b40bf45430320a4f30add05c8e3e16dd56c52d65f7abe475189564bf2b1ac00000000");
@@ -179,21 +180,22 @@ namespace DSW.HDWallet.AltCoins
                 PowNoRetargeting = true,
                 RuleChangeActivationThreshold = 1916,
                 MinerConfirmationWindow = 2016,
-                ConsensusFactory = PEPSCoinConsensusFactory.Instance,
+                ConsensusFactory = FlitsConsensusFactory.Instance,
                 SupportSegwit = false
             })
-               .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 55 })
-                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 92 })
-                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 117 })
-                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x04, 0x2e, 0x21, 0x71 })
-                .SetBase58Bytes(Base58Type.EXT_SECRET_KEY, new byte[] { 0x08, 0x25, 0x39, 0x2F })
-                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("PEPSCoin"))
-                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("PEPSCoin"))
+                .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 36 })
+                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 48 })
+                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 66 })
+                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x02, 0x2D, 0x25, 0x33 })
+                .SetBase58Bytes(Base58Type.EXT_SECRET_KEY, new byte[] { 0x02, 0x21, 0x31, 0x2B })
+                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("Flits"))
+                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("Flits"))
                 .SetMagic(0x191643a0)
-                .SetPort(15252)
+                .SetPort(32972)
+                .SetRPCPort(32973)
                 .SetMaxP2PVersion(70920)
-                .SetName("PEPSCoin-reg")
-                .AddAlias("PEPSCoin-regtest")
+                .SetName("Flits-reg")
+                .AddAlias("Flits-regtest")
                 .AddDNSSeeds(new DNSSeedData[0])
                 .AddSeeds(new NetworkAddress[0])
                 //No regtest at the moment
