+++ DSW.HDWallet/Infrastructure/Coins/Jackpot.cs
@@ -4,40 +4,40 @@ using NBitcoin.Crypto;
 using NBitcoin.DataEncoders;
 using NBitcoin.Protocol;
 
-namespace DSW.HDWallet.AltCoins
+namespace DSW.HDWallet.Infrastructure.Coins
 {
 
-    public class OneWorldCoin : NetworkSetBase
+    public class Jackpot : NetworkSetBase
     {
-        public static OneWorldCoin Instance { get; } = new OneWorldCoin();
+        public static Jackpot Instance { get; } = new Jackpot();
 
-        public override string CryptoCode => "OWO";
+        public override string CryptoCode => "777";
 
-        private OneWorldCoin()
+        private Jackpot()
         {
         }
 
-        public class OneWorldCoinConsensusFactory : ConsensusFactory
+        public class JackpotConsensusFactory : ConsensusFactory
         {
-            private OneWorldCoinConsensusFactory()
+            private JackpotConsensusFactory()
             {
             }
 
-            public static OneWorldCoinConsensusFactory Instance { get; } = new OneWorldCoinConsensusFactory();
+            public static JackpotConsensusFactory Instance { get; } = new JackpotConsensusFactory();
 
             public override BlockHeader CreateBlockHeader()
             {
-                return new OneWorldCoinBlockHeader();
+                return new JackpotBlockHeader();
             }
 
             public override Block CreateBlock()
             {
-                return new OneWorldCoinBlock(new OneWorldCoinBlockHeader());
+                return new JackpotBlock(new JackpotBlockHeader());
             }
         }
 
 #pragma warning disable CS0618 // Type or member is obsolete
-        public class OneWorldCoinBlockHeader : BlockHeader
+        public class JackpotBlockHeader : BlockHeader
         {
             private static byte[] CalculateHash(byte[] data, int offset, int count)
             {
@@ -52,9 +52,9 @@ namespace DSW.HDWallet.AltCoins
             }
         }
 
-        public class OneWorldCoinBlock : Block
+        public class JackpotBlock : Block
         {
-            public OneWorldCoinBlock(OneWorldCoinBlockHeader h) : base(h)
+            public JackpotBlock(JackpotBlockHeader h) : base(h)
             {
             }
 
@@ -67,7 +67,7 @@ namespace DSW.HDWallet.AltCoins
 
         protected override void PostInit()
         {
-            RegisterDefaultCookiePath("OWO");
+            RegisterDefaultCookiePath("777");
         }
 
         protected override NetworkBuilder CreateMainnet()
@@ -79,7 +79,7 @@ namespace DSW.HDWallet.AltCoins
                 MajorityEnforceBlockUpgrade = 8100,
                 MajorityRejectBlockOutdated = 10260,
                 MajorityWindow = 10800,
-                BIP34Hash = new uint256("00000eef0583695d6da23a78bab1c39939bbb54cf9bd5f0d4881c8eef364cd26"),
+                BIP34Hash = new uint256("00000366a6d89ece72f3481f4c81e813f17c4512d2cce519c9d869d8eaebf71d"),
                 PowLimit = new Target(0 >> 1),
                 MinimumChainWork = new uint256("000000000000000000000000000000000000000000000000c2ba8ca4fb1f06cb"),
                 PowTargetTimespan = TimeSpan.FromSeconds(24 * 60 * 60),
@@ -89,28 +89,28 @@ namespace DSW.HDWallet.AltCoins
                 PowNoRetargeting = false,
                 RuleChangeActivationThreshold = 1916,
                 MinerConfirmationWindow = 2016,
-                ConsensusFactory = OneWorldCoinConsensusFactory.Instance,
+                ConsensusFactory = JackpotConsensusFactory.Instance,
                 SupportSegwit = false,
-                CoinType = 846
+                CoinType = 833
             })
-                .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 115 })
-                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 112 })
-                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 92 })
-                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x02, 0x2D, 0x25, 0x33 })
-                .SetBase58Bytes(Base58Type.EXT_SECRET_KEY, new byte[] { 0x02, 0x21, 0x31, 0x2B })
-                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("OneWorldCoin"))
-                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("OneWorldCoin"))
+                .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 15 })
+                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 16 })
+                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 43 })
+                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x02, 0x2D, 0x25, 0x73 })
+                .SetBase58Bytes(Base58Type.EXT_SECRET_KEY, new byte[] { 0x02, 0x21, 0x31, 0x2B})
+                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("Jackpot"))
+                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("Jackpot"))
                 .SetMagic(0x191643a0)
-                .SetPort(32112)
-                .SetRPCPort(32113)
+                .SetPort(27771)
+                .SetRPCPort(27772)
                 .SetMaxP2PVersion(70920)
-                .SetName("OneWorldCoin-main")
-                .AddAlias("OneWorldCoin-mainnet")
+                .SetName("Jackpot-main")
+                .AddAlias("Jackpot-mainnet")
                 .AddDNSSeeds(new[]
                 {
-                    new DNSSeedData("seed1", "seed1.owocoin.net"),
-                    new DNSSeedData("seed2", "seed2.owocoin.net"),
-                    new DNSSeedData("seed3", "seed3.owocoin.net")
+                    new DNSSeedData("seed1", "seed1.777coin.win"),
+                    new DNSSeedData("seed2", "seed2.777coin.win"),
+                    new DNSSeedData("seed3", "seed3.777coin.win")
                 })
                 .AddSeeds(new NetworkAddress[0])
                 .SetGenesis("01000000000000000000000000000000000000000000000000000000000000000000000014e427b75837280517873799a954e87b8b0484f3f1df927888a0ff4fd3a0c9f7bb2eac56f0ff0f1edfa624000101000000010000000000000000000000000000000000000000000000000000000000000000ffffffff8604ffff001d01044c7d323031372d30392d32312032323a30313a3034203a20426974636f696e20426c6f636b204861736820666f722048656967687420343836333832203a2030303030303030303030303030303030303039326431356535623365366538323639333938613834613630616535613264626434653766343331313939643033ffffffff0100ba1dd205000000434104c10e83b2703ccf322f7dbd62dd5855ac7c10bd055814ce121ba32607d573b8810c02c0582aed05b4deb9c4b77b26d92428c61256cd42774babea0a073b2ed0c9ac00000000");
@@ -137,23 +137,23 @@ namespace DSW.HDWallet.AltCoins
                 PowNoRetargeting = false,
                 RuleChangeActivationThreshold = 1512,
                 MinerConfirmationWindow = 2016,
-                ConsensusFactory = OneWorldCoinConsensusFactory.Instance,
+                ConsensusFactory = JackpotConsensusFactory.Instance,
                 SupportSegwit = false,
                 CoinType = 1
             })
-                .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 115 })
-                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 112 })
-                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 92 })
-                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x02, 0x2D, 0x25, 0x33 })
+                .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 15 })
+                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 16 })
+                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 43 })
+                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x02, 0x2D, 0x25, 0x73 })
                 .SetBase58Bytes(Base58Type.EXT_SECRET_KEY, new byte[] { 0x02, 0x21, 0x31, 0x2B })
-                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("OneWorldCoin"))
-                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("OneWorldCoin"))
+                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("Jackpot"))
+                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("Jackpot"))
                 .SetMagic(0x191643a0)
-                .SetPort(32112)
-                .SetRPCPort(32113)
+                .SetPort(27771)
+                .SetRPCPort(27772)
                 .SetMaxP2PVersion(70920)
-                .SetName("OneWorldCoin-test")
-                .AddAlias("OneWorldCoin-testnet")
+                .SetName("Jackpot-test")
+                .AddAlias("Jackpot-testnet")
                 .AddSeeds(new NetworkAddress[0])
                 //testnet down for now
                 .SetGenesis("0100000000000000000000000000000000000000000000000000000000000000000000008c5b00d67050180b3a90addb9cd1aabbb3dd79ce20fc071d428ce374581b3f7cde30df5cf0ff0f1e1a1754000101000000010000000000000000000000000000000000000000000000000000000000000000ffffffff3b04ffff001d010433446f676543617368205265706f7765726564204c61756e6368202d20616b736861796e65787573202d204c6971756964333639ffffffff0100000000000000004341047a7df379bd5e6b93b164968c10fcbb141ecb3c6dc1a5e181c2a62328405cf82311dd5b40bf45430320a4f30add05c8e3e16dd56c52d65f7abe475189564bf2b1ac00000000");
@@ -180,22 +180,22 @@ namespace DSW.HDWallet.AltCoins
                 PowNoRetargeting = true,
                 RuleChangeActivationThreshold = 1916,
                 MinerConfirmationWindow = 2016,
-                ConsensusFactory = OneWorldCoinConsensusFactory.Instance,
+                ConsensusFactory = JackpotConsensusFactory.Instance,
                 SupportSegwit = false
             })
-                .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 115 })
-                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 112 })
-                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 92 })
-                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x02, 0x2D, 0x25, 0x33 })
+                .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { 15 })
+                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { 16 })
+                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { 43 })
+                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { 0x02, 0x2D, 0x25, 0x73 })
                 .SetBase58Bytes(Base58Type.EXT_SECRET_KEY, new byte[] { 0x02, 0x21, 0x31, 0x2B })
-                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("OneWorldCoin"))
-                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("OneWorldCoin"))
+                .SetBech32(Bech32Type.WITNESS_PUBKEY_ADDRESS, Encoders.Bech32("Jackpot"))
+                .SetBech32(Bech32Type.WITNESS_SCRIPT_ADDRESS, Encoders.Bech32("Jackpot"))
                 .SetMagic(0x191643a0)
-                .SetPort(32112)
-                .SetRPCPort(32113)
+                .SetPort(27771)
+                .SetRPCPort(27772)
                 .SetMaxP2PVersion(70920)
-                .SetName("OneWorldCoin-reg")
-                .AddAlias("OneWorldCoin-regtest")
+                .SetName("Jackpot-reg")
+                .AddAlias("Jackpot-regtest")
                 .AddDNSSeeds(new DNSSeedData[0])
                 .AddSeeds(new NetworkAddress[0])
                 //No regtest at the moment
