+++ DSW.HDWallet/Infrastructure/WalletRepository.cs
@@ -1,28 +1,35 @@
-﻿using DSW.HDWallet.Domain.Wallets;
+﻿using DSW.HDWallet.Domain.Coins;
+using DSW.HDWallet.Domain.Wallets;
 using NBitcoin;
 
 namespace DSW.HDWallet.Infrastructure
 {
     public class WalletRepository : IWalletRepository
     {
-        // TO DO - Configurar a injeção de dependência para fornecer a instância de Network quando você instancia a classe WalletRepository
-
-        /*
-        private readonly Network _network;
-
-        public WalletRepository(Network network)
+        public Wallet Create(Mnemonic mnemo, CoinType coinType)
         {
-            _network = network ?? throw new ArgumentNullException(nameof(network));
-        }
-        */
-
-        public Wallet Create(Mnemonic mnemo)
-        {
-            BitcoinAddress address = GetAddress(mnemo);
+            Network network = CoinNetwork.GetMainnet(coinType);
+            BitcoinAddress address = GetAddress(coinType, mnemo);
 
             var wallet = new Wallet
             {
-                MasterKey = mnemo.DeriveExtKey().ToString(Network.Main),
+                MasterKey = mnemo.DeriveExtKey().ToString(network),
+                Address = address.ToNetwork(network).ToString(),
+                SecretWords = mnemo.ToString(),
+                SecrectWordsArray = mnemo.Words
+            };
+
+            return wallet;
+        }
+
+        public Wallet CreateWithPassword(CoinType coinType, Mnemonic mnemo, string? password = null)
+        {
+            Network network = CoinNetwork.GetMainnet(coinType);
+            BitcoinAddress address = GetAddress(coinType, mnemo, password);
+
+            var wallet = new Wallet
+            {
+                MasterKey = mnemo.DeriveExtKey().ToString(network),
                 Address = address.ToString(),
                 SecretWords = mnemo.ToString(),
                 SecrectWordsArray = mnemo.Words
@@ -31,38 +38,26 @@ namespace DSW.HDWallet.Infrastructure
             return wallet;
         }
 
-        public Wallet CreateWithPassword(Mnemonic mnemo, string? password = null)
-        {            
-            BitcoinAddress address = GetAddress(mnemo, password);
-
-            var wallet = new Wallet
-            {
-                MasterKey = mnemo.DeriveExtKey().ToString(Network.Main),
-                Address = address.ToString(),
-                SecretWords = mnemo.ToString(),
-                SecrectWordsArray = mnemo.Words
-            };
-
-            return wallet;
+        public BitcoinAddress Recover(CoinType coinType, Mnemonic mnemo, string? password = null)
+        {
+            return GetAddress(coinType, mnemo, password);
         }
 
-        public BitcoinAddress Recover(Mnemonic mnemo, string? password = null)
+        private static BitcoinAddress GetAddress(CoinType coinType, Mnemonic mnemo, string? password = null)
         {
-            return GetAddress(mnemo, password);
-        }
-
-        private static BitcoinAddress GetAddress(Mnemonic mnemo, string? password = null)
-        {
+            Network network = CoinNetwork.GetMainnet(coinType);
             ExtKey masterKey = string.IsNullOrEmpty(password) ? mnemo.DeriveExtKey() : mnemo.DeriveExtKey(password);
             ExtPubKey masterPubKey = masterKey.Neuter();
-            BitcoinAddress address = masterPubKey.PubKey.GetAddress(ScriptPubKeyType.Legacy, Network.Main);
+            BitcoinAddress address = masterPubKey.PubKey.GetAddress(ScriptPubKeyType.Legacy, network);
 
             return address;
         }
 
-        public BitcoinExtKey CreateDeriveKey(ExtKey masterKey, KeyPath keyPath)
+        public BitcoinExtKey CreateDeriveKey(CoinType coinType, ExtKey masterKey, KeyPath keyPath)
         {
-            return new BitcoinExtKey(masterKey.Derive(keyPath), Network.Main);
+            Network network = CoinNetwork.GetMainnet(coinType);
+
+            return new BitcoinExtKey(masterKey.Derive(keyPath), network);
         }
     }
 }
